applied PR review changes
