--- apps/journeys/src/components/blocks/Trigger/index.ts
@@ -1,2 +1,0 @@
-export { Trigger } from './Trigger'
-export { TRIGGER_FIELDS } from './triggerFields'
