+++ apps/journeys/src/components/blocks/Video/Trigger/index.ts
@@ -1,0 +1,2 @@
+export { Trigger } from './Trigger'
+export { VIDEO_TRIGGER_FIELDS } from './triggerFields'
