+++ apps/journeys/src/components/blocks/Video/Trigger/triggerFields.ts
@@ -1,9 +1,9 @@
 import { gql } from '@apollo/client'
-import { ACTION_FIELDS } from '../../../libs/action'
+import { ACTION_FIELDS } from '../../../../libs/action'
 
-export const TRIGGER_FIELDS = gql`
+export const VIDEO_TRIGGER_FIELDS = gql`
   ${ACTION_FIELDS}
-  fragment TriggerFields on VideoTriggerBlock {
+  fragment VideoTriggerFields on VideoTriggerBlock {
     id
     parentBlockId
     triggerStart
