+++ apps/journeys/src/components/blocks/Video/index.ts
@@ -1,2 +1,4 @@
 export { Video, VIDEO_RESPONSE_CREATE } from './Video'
 export { VIDEO_FIELDS } from './videoFields'
+export { Trigger } from './Trigger'
+export { VIDEO_TRIGGER_FIELDS } from './Trigger'
