+++ apps/journeys/src/components/blocks/index.ts
@@ -7,6 +7,5 @@ export { RadioOption, RADIO_OPTION_FIELDS } from './RadioOption'
 export { RadioQuestion, RADIO_QUESTION_FIELDS } from './RadioQuestion'
 export { SignUp, SIGN_UP_FIELDS } from './SignUp'
 export { Step, STEP_FIELDS } from './Step'
-export { TRIGGER_FIELDS } from './Trigger'
 export { Typography, TYPOGRAPHY_FIELDS } from './Typography'
-export { Video, VIDEO_FIELDS } from './Video'
+export { Video, VIDEO_FIELDS, VIDEO_TRIGGER_FIELDS } from './Video'
